creación modulos controler y archivo index.js
