+++ src/lib/index.js
@@ -1,31 +1,26 @@
-// aqui exportaras las funciones que necesites
 
-//export const myFunction = () => {
-  // aqui tu codigo
-  //console.log('Hola mundo!');
-//};
+    // Import the functions you need from the SDKs you need
+    import { initializeApp } from "https://www.gstatic.com/firebasejs/9.9.1/firebase-app.js";
+    import { getAnalytics } from "https://www.gstatic.com/firebasejs/9.9.1/firebase-analytics.js";
+    // TODO: Add SDKs for Firebase products that you want to use
+    // https://firebase.google.com/docs/web/setup#available-libraries
+  
+    // Your web app's Firebase configuration
+    // For Firebase JS SDK v7.20.0 and later, measurementId is optional
+    const firebaseConfig = {
+      apiKey: "AIzaSyCrjTLw9QLhoJzC7qHJiKwFu3DRdtj1xgY",
+      authDomain: "laberintolector-5688a.firebaseapp.com",
+      projectId: "laberintolector-5688a",
+      storageBucket: "laberintolector-5688a.appspot.com",
+      messagingSenderId: "800505099476",
+      appId: "1:800505099476:web:e8781a0717d6a00bbbf741",
+      measurementId: "G-4MXS3Q2K9Q"
+    };
+  
+    // Initialize Firebase
+    const app = initializeApp(firebaseConfig);
+    const analytics = getAnalytics(app);
 
-import {components} from '../view/index.js'
+    
 
-const changeView = (route)=> {
-  const id = route.split('/')[1];
-  const container = document.getElementById('container')
-  container.innerHTML = '';
-  switch (route){
-    case '':
-    case '#':
-    case '#/':
-      { return container.appendChild(components.home())}
-    case '#/Register':
-      { return container.appendChild(components.register())}
-    case '#/Acces':
-      { return container.appendChild(components.acces())}
-    case '#/Post':
-      { return container.appendChild(components.post())}
-    default:
-      return container.appendChild(components.different())
-  }
-  console.log(route)
-}
 
-export {changeView}
